genral fixes (names, variables, etc) on Controller, DAO, Model & WS

+++ src/main/java/gr/teicm/offersaroundws/controller/OfferstempController.java
@@ -13,7 +13,7 @@ public class OfferstempController {
       this.view = view;
    }
     
-    public Integer getBusinessId(){
+    public Short getBusinessId(){
        return model.getBusinessId();
    }
